admin_login_&__menu updates
After much toil, I have decided it's easier to keep blood test report info in the donor file. Creating an extra file for this is difficult to keep up to date. It will work out by adding an extra element to the vector which will remain hidden on input and information viewing